updates AdCard to disable cta for MyAd
