+++ apps/frontend/components/sections/index.ts
@@ -1,1 +1,2 @@
 export { default as Market } from './Market';
+export { default as P2PAds } from './P2PAds';
